Bugfix: fix the python BlockManager and clean dsa.py to rm those unused host tensor.
